backend/requirements.txt
numpy
opencv-python
pillow
torch
torchvision
piexif
flask
flask-cors
